requirements.txt
python-telegram-bot==21.0
notion-client==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
APScheduler==3.10.4
httpx==0.27.0
anthropic==0.39.0
requests==2.31.0
twilio==9.0.0
agentmail
pdfplumber==0.11.4
openpyxl==3.1.5
